Add support for puppetdb behind basic auth
We have a moderately unique puppet setup (though remarkably close to
github's setup, in some ways). We run puppetdb behind a proxy that is
secured with basic auth (and we will eventually be submitting patches
upstream to puppet). This commit allows octocatalog-diff to parse basic
auth username, and password from a URL.