Normalize file resource titles
We've noticed a problem at Braintree along the lines of:

  file { '/foo':
    ensure => directory,
  }

  file { '/bar':
    ensure => directory,
    require => File['/foo/'],
  }

This catalog compiles and is valid, but fails validation in octocatalog-diff.
It seems that puppet itself attempts to normalize file resource titles (and indeed
can normalize any type of resource title, given that a methodology exists in the
resource type). See:
https://github.com/puppetlabs/puppet/blob/4.10.x/lib/puppet/type/file.rb#L42

This commit attmepts to mimic that (and lifts the regex directly from the 4.10.x
branch of puppet, because it doesn't matter in puppet 5 since reference validation
is built-in).