Add support for PE package inventory facts
When an agent gathers facts in a Puppet Enterprise environment with
the PE-specific package inventory feature turned on, the package
inventory is returned as part of the fact upload and then that
package data is available in a PuppetDB endpoint independent of the
rest of the facter data.

You should not generally need to include the package inventory data
for Octocatalog-Diff testing, but if you have a specific reason to
want it included to more closely match what occurs in production
this addition allows it to be included.